cli/requirements.txt
azure-storage-blob==12.19.0
python-dotenv==1.0.0
pathspec==0.12.1
click==8.1.7
colorama==0.4.6
tqdm==4.66.1